updated: DRY test
