+++ tests/unit/onlyFirstElemOption.test.js
@@ -1,45 +1,8 @@
-const { join } = require("path");
-const { fork } = require("child_process");
+const { setupTest } = require("../testHelper");
 
-describe("onlyFirstElem option", () => {
-  const testProcessPath = join(
-    __dirname,
-    ".",
-    "onlyFirstElemOption.test.process.js"
-  );
+describe("onlyFirstElem option",  () => {
 
-  let workerProcess;
-  let testCaseNum = 0;
-
-  function createWorkerDataPromise() {
-    return new Promise((resolve, reject) => {
-      let data = "";
-
-      workerProcess.stderr.on("data", (chunk) => {
-        data += chunk;
-      });
-
-      workerProcess.stderr.on("end", () => {
-        resolve(data);
-      });
-
-      workerProcess.stderr.on("error", reject);
-    });
-  }
-
-  function createMessagePromise() {
-    return new Promise((resolve, reject) => {
-      workerProcess.on("error", reject);
-
-      //once | on - either way, dependent on worker
-      workerProcess.once("message", (message) => {
-        if (message?.error) {
-          reject(new Error(message.error));
-        }
-        resolve(message);
-      });
-    });
-  }
+  const {createMessagePromise, createWorkerDataPromise} =  setupTest("unit", "onlyFirstElemOption.test.process.js")
 
   /**
    *
@@ -69,18 +32,6 @@ describe("onlyFirstElem option", () => {
 
     return { length, discolouredResult };
   }
-
-  beforeEach(() => {
-    workerProcess = fork(testProcessPath, {
-      stdio: [0, "pipe", "pipe", "ipc"],
-    });
-
-    testCaseNum++;
-
-    workerProcess.send(testCaseNum);
-
-    workerProcess.stderr.pipe(process.stderr, { end: false });
-  });
 
   describe("1) given n number of nested array functions", () => {
     it("should log only the first element; the outer arrays are ignored recursively", async () => {
